refactor(webapp): pay down boy-scout debt in packages/webapp/src/fs/mount-picker-popup.ts
Move the unified picker launcher into fs/ (with base/ logger + runtime-env)
so mount helpers no longer import up into shell/ or core/, leave a shell
re-export for device pickers, name DevicePickerInfo / Chrome window create
options, and ratchet both baselines.
