+++ packages/dev-tools/tools/layer-back-edge-baseline.json
@@ -1,11 +1,9 @@
 {
   "packages/webapp/src/cdp/har-recorder.ts": 1,
-  "packages/webapp/src/fs/mount-picker-popup.ts": 2,
   "packages/webapp/src/fs/mount/backend-local.ts": 2,
   "packages/webapp/src/shell/bsh-watchdog.ts": 3,
   "packages/webapp/src/shell/supplemental-commands/hid-command.ts": 1,
   "packages/webapp/src/shell/supplemental-commands/host-command.ts": 7,
-  "packages/webapp/src/shell/supplemental-commands/picker-popup.ts": 1,
   "packages/webapp/src/shell/supplemental-commands/playwright/handlers/interaction.ts": 1,
   "packages/webapp/src/shell/supplemental-commands/playwright/handlers/routing.ts": 2,
   "packages/webapp/src/shell/supplemental-commands/playwright/snapshot.ts": 3,
