+++ packages/dev-tools/tools/record-string-unknown-baseline.json
@@ -43,7 +43,6 @@
   "packages/webapp/src/scoops/tray-leader/oauth-cdp-login.ts": 1,
   "packages/webapp/src/scoops/tray-leader/tab-teleport.ts": 3,
   "packages/webapp/src/shell/bsh-watchdog.ts": 1,
-  "packages/webapp/src/shell/supplemental-commands/picker-popup.ts": 3,
   "packages/webapp/src/shell/supplemental-commands/playwright/teleport.ts": 5,
   "packages/webapp/src/shell/supplemental-commands/secret-backends.ts": 3,
   "packages/webapp/src/shell/supplemental-commands/tsc-command.ts": 4,
